requirements.txt
alembic==1.14.0
bcrypt==4.2.1
black==24.10.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Reuploaded==1.4.0
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.2
fonttools==4.55.8
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mypy-extensions==1.0.0
numpy==2.2.2
numpy-stl==3.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pyglet==2.1.3
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-utils==3.9.1
pytz==2025.2
qrcode==8.1
razorpay==1.4.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==76.0.0
six==1.17.0
SQLAlchemy==2.0.36
threadpoolctl==3.6.0
trimesh==4.6.8
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
virtualenv==20.30.0
vtk==9.4.2
Werkzeug==3.1.3
WTForms==3.2.1
